# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "review_ducktective",
    "description": "Автоматическое code review на LLM с RAG по кодовой базе",
    "version": "0.1.0"
  },
  "paths": {
...
    "/repositories/{repository_id}/reviews": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Start Review",
        "operationId": "start_review_repositories__repository_id__reviews_post",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "List Reviews",
        "operationId": "list_reviews_repositories__repository_id__reviews_get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewRunSummary"
                  },
                  "title": "Response List Reviews Repositories  Repository Id  Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/{run_id}": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Get Review",
        "operationId": "get_review_reviews__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "reviews"
        ],
        "summary": "Delete Review",
        "operationId": "delete_review_reviews__run_id__delete",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/{run_id}/run": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Enqueue Review",
        "description": "Ставит прогон в очередь.\n\nРевью занимает минуты, поэтому HTTP-запрос не ждёт его завершения: клиент\nопрашивает статус через GET.",
        "operationId": "enqueue_review_reviews__run_id__run_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/{run_id}/findings/{finding_id}/feedback": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Submit Feedback",
        "description": "Фиксирует оценку находки.\n\nОтметки накапливаются как размеченная выборка для измерения качества.",
        "operationId": "submit_feedback_reviews__run_id__findings__finding_id__feedback_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitFeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/{run_id}/files/{file_id}/patch": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Get File Patch",
        "operationId": "get_file_patch_reviews__run_id__files__file_id__patch_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilePatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/reviews/{run_id}/files/{file_id}/content": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Get File Context",
        "operationId": "get_file_context_reviews__run_id__files__file_id__content_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          },
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "start_line",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Start Line"
            }
          },
          {
            "name": "end_line",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "End Line"
            }
          },
          {
            "name": "side",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/DiffSide",
              "default": "new"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChangeType": {
        "type": "string",
        "enum": [
          "added",
          "modified",
          "deleted",
          "renamed"
        ],
        "title": "ChangeType"
      },
      "DiffSide": {
        "type": "string",
        "enum": [
          "old",
          "new"
        ],
        "title": "DiffSide",
        "description": "Сторона диффа: находка может относиться к удалённому или к новому коду."
      },
...
      "FeedbackResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "verdict": {
            "$ref": "#/components/schemas/FeedbackVerdict"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "verdict",
          "comment",
          "created_at"
        ],
        "title": "FeedbackResponse"
      },
      "FeedbackVerdict": {
        "type": "string",
        "enum": [
          "useful",
          "false_positive",
          "wontfix"
        ],
        "title": "FeedbackVerdict",
        "description": "Оценка находки человеком. Источник данных для eval-набора."
      },
      "FileContextResponse": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "side": {
            "$ref": "#/components/schemas/DiffSide"
          },
          "commit_sha": {
            "type": "string",
            "title": "Commit Sha"
          },
          "start_line": {
            "type": "integer",
            "title": "Start Line"
          },
          "end_line": {
            "type": "integer",
            "title": "End Line"
          },
          "total_lines": {
            "type": "integer",
            "title": "Total Lines"
          },
          "lines": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Lines"
          }
        },
        "type": "object",
        "required": [
          "path",
          "side",
          "commit_sha",
          "start_line",
          "end_line",
          "total_lines",
          "lines"
        ],
        "title": "FileContextResponse"
      },
      "FilePatchResponse": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "previous_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Path"
          },
          "change_type": {
            "$ref": "#/components/schemas/ChangeType"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "added_lines": {
            "type": "integer",
            "title": "Added Lines"
          },
          "removed_lines": {
            "type": "integer",
            "title": "Removed Lines"
          },
          "is_too_large": {
            "type": "boolean",
            "title": "Is Too Large"
          },
          "patch_size_bytes": {
            "type": "integer",
            "title": "Patch Size Bytes"
          },
          "patch": {
            "type": "string",
            "title": "Patch"
          }
        },
        "type": "object",
        "required": [
          "path",
          "previous_path",
          "change_type",
          "language",
          "added_lines",
          "removed_lines",
          "is_too_large",
          "patch_size_bytes",
          "patch"
        ],
        "title": "FilePatchResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewFileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "previous_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Path"
          },
          "change_type": {
            "$ref": "#/components/schemas/ChangeType"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "added_lines": {
            "type": "integer",
            "title": "Added Lines"
          },
          "removed_lines": {
            "type": "integer",
            "title": "Removed Lines"
          },
          "is_too_large": {
            "type": "boolean",
            "title": "Is Too Large"
          },
          "hunks": {
            "items": {
              "$ref": "#/components/schemas/HunkResponse"
            },
            "type": "array",
            "title": "Hunks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "path",
          "previous_path",
          "change_type",
          "language",
          "added_lines",
          "removed_lines",
          "is_too_large",
          "hunks"
        ],
        "title": "ReviewFileResponse"
      },
      "ReviewRunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id"
          },
          "source": {
            "$ref": "#/components/schemas/ReviewSource"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "base_sha": {
            "type": "string",
            "title": "Base Sha"
          },
          "head_sha": {
            "type": "string",
            "title": "Head Sha"
          },
          "totals": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Totals"
          },
          "failure_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ReviewFileResponse"
            },
            "type": "array",
            "title": "Files"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingResponse"
            },
            "type": "array",
            "title": "Findings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "repository_id",
          "source",
          "status",
          "base_sha",
          "head_sha",
          "totals",
          "failure_reason",
          "created_at",
          "started_at",
          "finished_at",
          "files",
          "findings"
        ],
        "title": "ReviewRunResponse"
      },
      "ReviewRunSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "base_sha": {
            "type": "string",
            "title": "Base Sha"
          },
          "head_sha": {
            "type": "string",
            "title": "Head Sha"
          },
          "totals": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Totals"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "changed_files": {
            "type": "integer",
            "title": "Changed Files"
          }
        },
        "type": "object",
        "required": [
          "id",
          "repository_id",
          "status",
          "base_sha",
          "head_sha",
          "totals",
          "created_at",
          "changed_files"
        ],
        "title": "ReviewRunSummary"
      },
      "ReviewSource": {
        "type": "string",
        "enum": [
          "pull_request",
          "local_diff",
          "upload"
        ],
        "title": "ReviewSource"
      },
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "queued",
          "indexing",
          "running",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "ReviewStatus"
      },
...
      "StartReviewRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "base": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Base"
          },
          "head": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Head",
            "default": "HEAD"
          },
          "source": {
            "$ref": "#/components/schemas/ReviewSource",
            "default": "local_diff"
          },
          "external_pull_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Pull Request Id"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "base"
        ],
        "title": "StartReviewRequest"
      },
      "SubmitFeedbackRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "verdict": {
            "$ref": "#/components/schemas/FeedbackVerdict"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "verdict"
        ],
        "title": "SubmitFeedbackRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}